I changed the data type of the ids

+++ src/main/java/com/matias/springjwt/repository/IRefreshTokenRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface IRefreshTokenRepository extends JpaRepository<RefreshTokenEntity, Long> {
+public interface IRefreshTokenRepository extends JpaRepository<RefreshTokenEntity, Integer> {
     Optional<RefreshTokenEntity> findByToken(String token);
 
     @Modifying
